requirements.txt
finnhub-python
confluent-kafka
pyspark
python-dotenv
apache-airflow
boto3
aws-msk-iam-sasl-signer-python

